test custom with scheduler fixture
This allows to add new schedulers in the future and have pytest
automatically create the tests for us. The should also catch more
bugs.
